move all files to src root
