+++ src/structs.rs
